backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
motor==3.6.0
python-dotenv==1.0.1
bcrypt==5.0.0
python-jose[cryptography]==3.3.0

